Decouple 8claw-specific pieces from shared chat core
- Move 8claw tool-result → card mapping out of lib/chat into
  plugins/app/toolResultBuilders.ts; lib/chat/toolResultPlugins.ts keeps
  only the generic injection mechanism
- Replace hardcoded 'Welcome to 8Claw' empty-state copy with a neutral
  default overridable via NEXT_PUBLIC_WELCOME_MESSAGE
